pyproject.toml
[project]
name = "fastai"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi[standard]>=0.135.2",
    "openai>=1.68.0",
    "pydantic-settings>=2.13.1",
    "pymupdf>=1.24.0",
    "python-docx>=1.1.2",
]

[dependency-groups]
dev = [
    "ruff>=0.15.8",
]
